requirements.txt
boto
docopt
six
sosreport
